Ejercicio 9 Finalizado y correcciones al Ejercicio 8

+++ src/Ejercicio8.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 void main() {
     Scanner teclado = new Scanner(System.in);
     System.out.print("Ingrese el monto a invertir: ");
-    double capital = teclado.nextLong();
+    double capital = teclado.nextDouble();
     System.out.print("Ingrese el interes anual: ");
-    double interes = teclado.nextInt();
+    double interes = teclado.nextDouble();
     System.out.print("Ingrese la cantidad de años: ");
     int anios = teclado.nextInt();
     for (int i = 1; i <= anios; i++) {
